+++ src/ui/gui_components/gui_status_bar.c
@@ -33,9 +33,6 @@ static int g_currentDisplayPercent = -1;
 
 typedef struct StatusBar {
     lv_obj_t *background;
-#if (WALLPAPER_ENABLE == 1)
-    lv_obj_t *wallPaper;
-#endif
     lv_obj_t *cont;
     lv_obj_t *walletIcon;
     lv_obj_t *walletNameLabel;
@@ -175,26 +172,9 @@ void GuiNvsBarSetWalletIcon(const void *src)
     lv_obj_align(g_guiStatusBar.walletIcon, LV_ALIGN_LEFT_MID, 26, 0);
 }
 
-void ShowWallPaper(bool enable)
-{
-#if (WALLPAPER_ENABLE == 1)
-    if (enable) {
-        lv_obj_clear_flag(g_guiStatusBar.wallPaper, LV_OBJ_FLAG_HIDDEN);
-    } else {
-        lv_obj_add_flag(g_guiStatusBar.wallPaper, LV_OBJ_FLAG_HIDDEN);
-    }
-#endif
-}
-
 void GuiStatusBarInit(void)
 {
-    g_guiStatusBar.background = GuiCreateContainer(
-                                    lv_obj_get_width(lv_scr_act()), lv_obj_get_height(lv_scr_act()));
-#if (WALLPAPER_ENABLE == 1)
-    g_guiStatusBar.wallPaper = GuiCreateImg(g_guiStatusBar.background, NULL);
-    lv_img_set_src(g_guiStatusBar.wallPaper, &imgDeepLayersVolume11);
-    ShowWallPaper(false);
-#endif
+    g_guiStatusBar.background = GuiCreateContainer(lv_obj_get_width(lv_scr_act()), lv_obj_get_height(lv_scr_act()));
     lv_obj_t *cont = GuiCreateContainerWithParent(g_guiStatusBar.background,
                      lv_obj_get_width(lv_scr_act()),
                      GUI_STATUS_BAR_HEIGHT);
