+++ src/ui/gui_views/multi/gui_multi_account_receive_view.c
@@ -1,4 +1,3 @@
-#ifndef BTC_ONLY
 #include "gui.h"
 #include "gui_obj.h"
 #include "gui_views.h"
@@ -50,4 +49,3 @@ GUI_VIEW g_multiAccountsReceiveView = {
     .optimization = false,
     .pEvtHandler = GuiMultiAccountsReceiveViewEventProcess,
 };
-#endif
