+++ src/ui/gui_views/multi/gui_standard_receive_view.c
@@ -1,4 +1,3 @@
-#ifndef BTC_ONLY
 #include "gui.h"
 #include "gui_obj.h"
 #include "gui_views.h"
@@ -55,4 +54,3 @@ GUI_VIEW g_standardReceiveView = {
     .optimization = false,
     .pEvtHandler = GuiStandardReceiveViewEventProcess,
 };
-#endif
