+++ src/ui/gui_views/multi/web3/gui_multi_path_coin_receive_view.c
@@ -1,4 +1,3 @@
-#ifndef BTC_ONLY
 #include "gui.h"
 #include "gui_obj.h"
 #include "gui_views.h"
@@ -50,4 +49,3 @@ GUI_VIEW g_multiPathCoinReceiveView = {
     .optimization = false,
     .pEvtHandler = GuiMultiPathCoinReceiveViewEventProcess,
 };
-#endif
