+++ src/ui/gui_widgets/btc_only/multi_sig/gui_multisig_read_sdcard_widgets.c
@@ -175,7 +175,5 @@ void GuiPSBTTransactionCheckFaild(PtrT_TransactionCheckResult result)
         break;
     }
     GuiModelTransactionCheckResultClear();
-#if BTC_ONLY
     FreePsbtUxtoMemory();
-#endif
 }
