+++ src/ui/gui_widgets/gui_connect_wallet_widgets.h
