+++ src/ui/gui_widgets/gui_lock_widgets.c
@@ -210,9 +210,6 @@ void GuiLockScreenTurnOn(void *param)
     if (*single == SIG_LOCK_VIEW_VERIFY_PIN || *single == SIG_LOCK_VIEW_SCREEN_GO_HOME_PASS) {
         GuiNvsBarSetWalletIcon(NULL);
         GuiNvsBarSetWalletName("");
-#ifdef BTC_ONLY
-        ShowWallPaper(false);
-#endif
     }
     lv_obj_clear_flag(g_pageWidget->page, LV_OBJ_FLAG_HIDDEN);
     // g_lockView.isActive = true;
