+++ src/ui/gui_widgets/gui_select_address_widgets.c
@@ -1,15 +1,15 @@
-#ifndef BTC_ONLY
 #include "gui_select_address_widgets.h"
 #include "gui_home_widgets.h"
 #include "account_public_info.h"
 #include "gui_page.h"
 #include "gui_obj.h"
 #include "gui_button.h"
-
+#include "gui_keyboard.h"
 
 #ifdef COMPILE_SIMULATOR
 #include "simulator_mock_define.h"
 #endif
+
 #define GENERAL_ADDRESS_INDEX_MAX                           999999999
 typedef struct {
     lv_obj_t *addressCountLabel;
@@ -50,178 +50,19 @@ static lv_obj_t *g_rightBtn;
 static SetSelectAddressIndexFunc g_setSelectIndexFunc;
 static lv_obj_t *g_confirmBtn;
 
-static JumpButtonWidgets_t g_standardJumpButtonWidgets ;
-
-
 static void SetCurrentSelectIndex(uint32_t selectIndex);
 static void UpdateConfirmBtn(void);
 static void BackHandler(lv_event_t *e);
 static void ConfirmHandler(lv_event_t *e);
+static void RefreshSwitchAccount(void);
+
+#ifdef WEB3_VERSION
+static JumpButtonWidgets_t g_standardJumpButtonWidgets;
+static void SetKeyboardValid(bool validation);
 static void JumpToAccountHandler(lv_event_t *e);
 static void CloseSwitchAddressHandler(lv_event_t *e);
 static void InputAddressIndexKeyboardHandler(lv_event_t *e);
-static void SetKeyboardValid(bool validation);
-static void RefreshSwitchAccount(void);
-
-static void CloseSwitchAddressHandler(lv_event_t *e)
-{
-    lv_obj_add_flag(g_standardJumpButtonWidgets.inputAccountCont, LV_OBJ_FLAG_HIDDEN);
-}
-
-
-static void SetKeyboardValid(bool validation)
-{
-    if (validation) {
-        if (lv_btnmatrix_has_btn_ctrl(g_standardJumpButtonWidgets.inputAccountKeyboard, 11, LV_BTNMATRIX_CTRL_DISABLED)) {
-            lv_btnmatrix_clear_btn_ctrl(g_standardJumpButtonWidgets.inputAccountKeyboard, 11, LV_BTNMATRIX_CTRL_DISABLED);
-        }
-        lv_btnmatrix_set_btn_ctrl(g_standardJumpButtonWidgets.inputAccountKeyboard, 11, LV_BTNMATRIX_CTRL_CHECKED);
-    } else {
-        if (lv_btnmatrix_has_btn_ctrl(g_standardJumpButtonWidgets.inputAccountKeyboard, 11, LV_BTNMATRIX_CTRL_CHECKED)) {
-            lv_btnmatrix_clear_btn_ctrl(g_standardJumpButtonWidgets.inputAccountKeyboard, 11, LV_BTNMATRIX_CTRL_CHECKED);
-        }
-        lv_btnmatrix_set_btn_ctrl(g_standardJumpButtonWidgets.inputAccountKeyboard, 11, LV_BTNMATRIX_CTRL_DISABLED);
-    }
-}
-
-static void InputAddressIndexKeyboardHandler(lv_event_t *e)
-{
-    lv_event_code_t code = lv_event_get_code(e);
-    lv_obj_t *obj = lv_event_get_target(e);
-    uint32_t id = lv_btnmatrix_get_selected_btn(obj);
-    lv_obj_draw_part_dsc_t *dsc;
-    const char *txt;
-    char input[16];
-    uint32_t len;
-    uint64_t longInt;
-
-    if (code == LV_EVENT_CLICKED) {
-        txt = lv_btnmatrix_get_btn_text(obj, id);
-        strcpy_s(input, sizeof(input), lv_label_get_text(g_standardJumpButtonWidgets.inputAccountLabel));
-        if (strcmp(txt, LV_SYMBOL_OK) == 0) {
-            if (g_standardJumpButtonWidgets.g_inputAccountValid) {
-                sscanf(input, "%u", &g_standardJumpButtonWidgets.g_inputTmpIndex);
-                g_showIndex = g_standardJumpButtonWidgets.g_inputTmpIndex / 5 * 5;
-                g_selectIndex = g_standardJumpButtonWidgets.g_inputTmpIndex;
-                RefreshSwitchAccount();
-                UpdateConfirmBtn();
-                lv_obj_add_flag(g_standardJumpButtonWidgets.inputAccountCont, LV_OBJ_FLAG_HIDDEN);
-                g_standardJumpButtonWidgets.g_inputAccountValid = false;
-            }
-        } else if (strcmp(txt, "-") == 0) {
-            len = strlen(input);
-            if (len >= 1) {
-                input[len - 1] = '\0';
-                lv_label_set_text(g_standardJumpButtonWidgets.inputAccountLabel, input);
-                lv_obj_add_flag(g_standardJumpButtonWidgets.overflowLabel, LV_OBJ_FLAG_HIDDEN);
-                if (strlen(input) >= 1) {
-                    g_standardJumpButtonWidgets.g_inputAccountValid = true;
-                } else {
-                    g_standardJumpButtonWidgets.g_inputAccountValid = false;
-                }
-            }
-        } else if (strlen(input) < 15) {
-            strcat(input, txt);
-            longInt = strtol(input, NULL, 10);
-            if (longInt >= GENERAL_ADDRESS_INDEX_MAX) {
-                input[9] = '\0';
-                lv_obj_clear_flag(g_standardJumpButtonWidgets.overflowLabel, LV_OBJ_FLAG_HIDDEN);
-            } else {
-                lv_obj_add_flag(g_standardJumpButtonWidgets.overflowLabel, LV_OBJ_FLAG_HIDDEN);
-            }
-            if (longInt > 0) {
-                if (input[0] == '0') {
-                    lv_label_set_text(g_standardJumpButtonWidgets.inputAccountLabel, input + 1);
-                } else {
-                    lv_label_set_text(g_standardJumpButtonWidgets.inputAccountLabel, input);
-                }
-            } else {
-                lv_label_set_text(g_standardJumpButtonWidgets.inputAccountLabel, "0");
-            }
-            g_standardJumpButtonWidgets.g_inputAccountValid = true;
-        } else {
-            g_standardJumpButtonWidgets.g_inputAccountValid = false;
-            printf("input to long\r\n");
-        }
-        SetKeyboardValid(g_standardJumpButtonWidgets.g_inputAccountValid);
-    } else if (code == LV_EVENT_DRAW_PART_BEGIN) {
-        dsc = lv_event_get_draw_part_dsc(e);
-        if (dsc->class_p == &lv_btnmatrix_class && dsc->type == LV_BTNMATRIX_DRAW_PART_BTN) {
-            /*Change the draw descriptor of the 12th button*/
-            if (dsc->id == 9) {
-                dsc->label_dsc->opa = LV_OPA_TRANSP;
-            } else if (dsc->id == 11) {
-                dsc->rect_dsc->bg_color = ORANGE_COLOR;
-                dsc->label_dsc->opa = LV_OPA_TRANSP;
-            } else {
-                dsc->rect_dsc->bg_color = DARK_GRAY_COLOR;
-            }
-        }
-    } else if (code == LV_EVENT_DRAW_PART_END) {
-        dsc = lv_event_get_draw_part_dsc(e);
-        /*When the button matrix draws the buttons...*/
-        if (dsc->class_p == &lv_btnmatrix_class && dsc->type == LV_BTNMATRIX_DRAW_PART_BTN) {
-            /*Add custom content to the 4th button when the button itself was drawn*/
-            if (dsc->id == 9 || dsc->id == 11) {
-                lv_img_header_t header;
-                lv_draw_img_dsc_t img_draw_dsc;
-                lv_area_t a;
-                const lv_img_dsc_t *imgDsc;
-                lv_res_t res;
-                imgDsc = dsc->id == 9 ? &imgBackspace : &imgCheck;
-                res = lv_img_decoder_get_info(imgDsc, &header);
-                if (res != LV_RES_OK)
-                    return;
-                a.x1 = dsc->draw_area->x1 + (lv_area_get_width(dsc->draw_area) - header.w) / 2;
-                a.x2 = a.x1 + header.w - 1;
-                a.y1 = dsc->draw_area->y1 + (lv_area_get_height(dsc->draw_area) - header.h) / 2;
-                a.y2 = a.y1 + header.h - 1;
-                lv_draw_img_dsc_init(&img_draw_dsc);
-                img_draw_dsc.recolor = lv_color_black();
-                if (lv_btnmatrix_get_selected_btn(obj) == dsc->id)
-                    img_draw_dsc.recolor_opa = LV_OPA_30;
-                lv_draw_img(dsc->draw_ctx, &img_draw_dsc, &a, imgDsc);
-            }
-        }
-    }
-}
-
-
-static void JumpToAccountHandler(lv_event_t *e)
-{
-    lv_obj_t *cont, *label, *line, *closeBtn;
-    static lv_point_t points[2] = {{0, 0}, {408, 0}};
-    g_standardJumpButtonWidgets.g_inputAccountValid = false;
-
-    g_standardJumpButtonWidgets.inputAccountCont = GuiCreateHintBox(530);
-    lv_obj_add_event_cb(lv_obj_get_child(g_standardJumpButtonWidgets.inputAccountCont, 0), CloseHintBoxHandler, LV_EVENT_CLICKED, &g_standardJumpButtonWidgets.inputAccountCont);
-    cont = g_standardJumpButtonWidgets.inputAccountCont;
-
-    label = GuiCreateNoticeLabel(cont, _("receive_btc_receive_change_address_title"));
-    lv_obj_align(label, LV_ALIGN_TOP_LEFT, 36, 30 + 270);
-    label = GuiCreateNoticeLabel(cont, "");
-    lv_label_set_text_fmt(label, "%s-", _("Address"));
-    lv_obj_align(label, LV_ALIGN_TOP_LEFT, 36, 108 + 270);
-    g_standardJumpButtonWidgets.inputAccountLabel = GuiCreateTextLabel(cont, "");
-    lv_obj_align(g_standardJumpButtonWidgets.inputAccountLabel, LV_ALIGN_TOP_LEFT, 38 + lv_obj_get_self_width(label), 108 + 270);
-    label = GuiCreateIllustrateLabel(cont, _("receive_btc_receive_change_address_limit"));
-    lv_obj_align(label, LV_ALIGN_TOP_LEFT, 36, 170 + 270);
-    lv_obj_set_style_text_color(label, RED_COLOR, LV_PART_MAIN);
-    lv_obj_add_flag(label, LV_OBJ_FLAG_HIDDEN);
-    g_standardJumpButtonWidgets.overflowLabel = label;
-
-    line = GuiCreateLine(cont, points, 2);
-    lv_obj_align(line, LV_ALIGN_TOP_MID, 0, 160 + 270);
-
-    lv_obj_t *keyboard = GuiCreateNumKeyboard(cont, InputAddressIndexKeyboardHandler, NUM_KEYBOARD_NORMAL, NULL);
-    lv_obj_align(keyboard, LV_ALIGN_BOTTOM_MID, 0, -2);
-    lv_obj_add_style(keyboard, &g_enterPressBtnmStyle, LV_STATE_PRESSED | LV_PART_ITEMS);
-    lv_btnmatrix_set_btn_ctrl(keyboard, 11, LV_BTNMATRIX_CTRL_DISABLED);
-    g_standardJumpButtonWidgets.inputAccountKeyboard = keyboard;
-
-    closeBtn = GuiCreateImgButton(cont, &imgClose, 40, CloseSwitchAddressHandler, NULL);
-    lv_obj_align(closeBtn, LV_ALIGN_TOP_RIGHT, -36, 27 + 270);
-}
+#endif
 
 static void SelectAddressHandler(lv_event_t *e)
 {
@@ -493,4 +334,162 @@ lv_obj_t *GuiCreateSelectAddressWidget(GuiChainCoinType chainCoinType, uint32_t 
     return cont;
 }
 
+#ifdef WEB3_VERSION
+static void JumpToAccountHandler(lv_event_t *e)
+{
+    lv_obj_t *cont, *label, *line, *closeBtn;
+    static lv_point_t points[2] = {{0, 0}, {408, 0}};
+    g_standardJumpButtonWidgets.g_inputAccountValid = false;
+
+    g_standardJumpButtonWidgets.inputAccountCont = GuiCreateHintBox(530);
+    lv_obj_add_event_cb(lv_obj_get_child(g_standardJumpButtonWidgets.inputAccountCont, 0), CloseHintBoxHandler, LV_EVENT_CLICKED, &g_standardJumpButtonWidgets.inputAccountCont);
+    cont = g_standardJumpButtonWidgets.inputAccountCont;
+
+    label = GuiCreateNoticeLabel(cont, _("receive_btc_receive_change_address_title"));
+    lv_obj_align(label, LV_ALIGN_TOP_LEFT, 36, 30 + 270);
+    label = GuiCreateNoticeLabel(cont, "");
+    lv_label_set_text_fmt(label, "%s-", _("Address"));
+    lv_obj_align(label, LV_ALIGN_TOP_LEFT, 36, 108 + 270);
+    g_standardJumpButtonWidgets.inputAccountLabel = GuiCreateTextLabel(cont, "");
+    lv_obj_align(g_standardJumpButtonWidgets.inputAccountLabel, LV_ALIGN_TOP_LEFT, 38 + lv_obj_get_self_width(label), 108 + 270);
+    label = GuiCreateIllustrateLabel(cont, _("receive_btc_receive_change_address_limit"));
+    lv_obj_align(label, LV_ALIGN_TOP_LEFT, 36, 170 + 270);
+    lv_obj_set_style_text_color(label, RED_COLOR, LV_PART_MAIN);
+    lv_obj_add_flag(label, LV_OBJ_FLAG_HIDDEN);
+    g_standardJumpButtonWidgets.overflowLabel = label;
+
+    line = GuiCreateLine(cont, points, 2);
+    lv_obj_align(line, LV_ALIGN_TOP_MID, 0, 160 + 270);
+
+    lv_obj_t *keyboard = GuiCreateNumKeyboard(cont, InputAddressIndexKeyboardHandler, NUM_KEYBOARD_NORMAL, NULL);
+    lv_obj_align(keyboard, LV_ALIGN_BOTTOM_MID, 0, -2);
+    lv_obj_add_style(keyboard, &g_enterPressBtnmStyle, LV_STATE_PRESSED | LV_PART_ITEMS);
+    lv_btnmatrix_set_btn_ctrl(keyboard, 11, LV_BTNMATRIX_CTRL_DISABLED);
+    g_standardJumpButtonWidgets.inputAccountKeyboard = keyboard;
+
+    closeBtn = GuiCreateImgButton(cont, &imgClose, 40, CloseSwitchAddressHandler, NULL);
+    lv_obj_align(closeBtn, LV_ALIGN_TOP_RIGHT, -36, 27 + 270);
+}
+
+static void CloseSwitchAddressHandler(lv_event_t *e)
+{
+    lv_obj_add_flag(g_standardJumpButtonWidgets.inputAccountCont, LV_OBJ_FLAG_HIDDEN);
+}
+
+static void InputAddressIndexKeyboardHandler(lv_event_t *e)
+{
+    lv_event_code_t code = lv_event_get_code(e);
+    lv_obj_t *obj = lv_event_get_target(e);
+    uint32_t id = lv_btnmatrix_get_selected_btn(obj);
+    lv_obj_draw_part_dsc_t *dsc;
+    const char *txt;
+    char input[16];
+    uint32_t len;
+    uint64_t longInt;
+
+    if (code == LV_EVENT_CLICKED) {
+        txt = lv_btnmatrix_get_btn_text(obj, id);
+        strcpy_s(input, sizeof(input), lv_label_get_text(g_standardJumpButtonWidgets.inputAccountLabel));
+        if (strcmp(txt, LV_SYMBOL_OK) == 0) {
+            if (g_standardJumpButtonWidgets.g_inputAccountValid) {
+                sscanf(input, "%u", &g_standardJumpButtonWidgets.g_inputTmpIndex);
+                g_showIndex = g_standardJumpButtonWidgets.g_inputTmpIndex / 5 * 5;
+                g_selectIndex = g_standardJumpButtonWidgets.g_inputTmpIndex;
+                RefreshSwitchAccount();
+                UpdateConfirmBtn();
+                lv_obj_add_flag(g_standardJumpButtonWidgets.inputAccountCont, LV_OBJ_FLAG_HIDDEN);
+                g_standardJumpButtonWidgets.g_inputAccountValid = false;
+            }
+        } else if (strcmp(txt, "-") == 0) {
+            len = strlen(input);
+            if (len >= 1) {
+                input[len - 1] = '\0';
+                lv_label_set_text(g_standardJumpButtonWidgets.inputAccountLabel, input);
+                lv_obj_add_flag(g_standardJumpButtonWidgets.overflowLabel, LV_OBJ_FLAG_HIDDEN);
+                if (strlen(input) >= 1) {
+                    g_standardJumpButtonWidgets.g_inputAccountValid = true;
+                } else {
+                    g_standardJumpButtonWidgets.g_inputAccountValid = false;
+                }
+            }
+        } else if (strlen(input) < 15) {
+            strcat(input, txt);
+            longInt = strtol(input, NULL, 10);
+            if (longInt >= GENERAL_ADDRESS_INDEX_MAX) {
+                input[9] = '\0';
+                lv_obj_clear_flag(g_standardJumpButtonWidgets.overflowLabel, LV_OBJ_FLAG_HIDDEN);
+            } else {
+                lv_obj_add_flag(g_standardJumpButtonWidgets.overflowLabel, LV_OBJ_FLAG_HIDDEN);
+            }
+            if (longInt > 0) {
+                if (input[0] == '0') {
+                    lv_label_set_text(g_standardJumpButtonWidgets.inputAccountLabel, input + 1);
+                } else {
+                    lv_label_set_text(g_standardJumpButtonWidgets.inputAccountLabel, input);
+                }
+            } else {
+                lv_label_set_text(g_standardJumpButtonWidgets.inputAccountLabel, "0");
+            }
+            g_standardJumpButtonWidgets.g_inputAccountValid = true;
+        } else {
+            g_standardJumpButtonWidgets.g_inputAccountValid = false;
+            printf("input to long\r\n");
+        }
+        SetKeyboardValid(g_standardJumpButtonWidgets.g_inputAccountValid);
+    } else if (code == LV_EVENT_DRAW_PART_BEGIN) {
+        dsc = lv_event_get_draw_part_dsc(e);
+        if (dsc->class_p == &lv_btnmatrix_class && dsc->type == LV_BTNMATRIX_DRAW_PART_BTN) {
+            /*Change the draw descriptor of the 12th button*/
+            if (dsc->id == 9) {
+                dsc->label_dsc->opa = LV_OPA_TRANSP;
+            } else if (dsc->id == 11) {
+                dsc->rect_dsc->bg_color = ORANGE_COLOR;
+                dsc->label_dsc->opa = LV_OPA_TRANSP;
+            } else {
+                dsc->rect_dsc->bg_color = DARK_GRAY_COLOR;
+            }
+        }
+    } else if (code == LV_EVENT_DRAW_PART_END) {
+        dsc = lv_event_get_draw_part_dsc(e);
+        /*When the button matrix draws the buttons...*/
+        if (dsc->class_p == &lv_btnmatrix_class && dsc->type == LV_BTNMATRIX_DRAW_PART_BTN) {
+            /*Add custom content to the 4th button when the button itself was drawn*/
+            if (dsc->id == 9 || dsc->id == 11) {
+                lv_img_header_t header;
+                lv_draw_img_dsc_t img_draw_dsc;
+                lv_area_t a;
+                const lv_img_dsc_t *imgDsc;
+                lv_res_t res;
+                imgDsc = dsc->id == 9 ? &imgBackspace : &imgCheck;
+                res = lv_img_decoder_get_info(imgDsc, &header);
+                if (res != LV_RES_OK)
+                    return;
+                a.x1 = dsc->draw_area->x1 + (lv_area_get_width(dsc->draw_area) - header.w) / 2;
+                a.x2 = a.x1 + header.w - 1;
+                a.y1 = dsc->draw_area->y1 + (lv_area_get_height(dsc->draw_area) - header.h) / 2;
+                a.y2 = a.y1 + header.h - 1;
+                lv_draw_img_dsc_init(&img_draw_dsc);
+                img_draw_dsc.recolor = lv_color_black();
+                if (lv_btnmatrix_get_selected_btn(obj) == dsc->id)
+                    img_draw_dsc.recolor_opa = LV_OPA_30;
+                lv_draw_img(dsc->draw_ctx, &img_draw_dsc, &a, imgDsc);
+            }
+        }
+    }
+}
+
+static void SetKeyboardValid(bool validation)
+{
+    if (validation) {
+        if (lv_btnmatrix_has_btn_ctrl(g_standardJumpButtonWidgets.inputAccountKeyboard, 11, LV_BTNMATRIX_CTRL_DISABLED)) {
+            lv_btnmatrix_clear_btn_ctrl(g_standardJumpButtonWidgets.inputAccountKeyboard, 11, LV_BTNMATRIX_CTRL_DISABLED);
+        }
+        lv_btnmatrix_set_btn_ctrl(g_standardJumpButtonWidgets.inputAccountKeyboard, 11, LV_BTNMATRIX_CTRL_CHECKED);
+    } else {
+        if (lv_btnmatrix_has_btn_ctrl(g_standardJumpButtonWidgets.inputAccountKeyboard, 11, LV_BTNMATRIX_CTRL_CHECKED)) {
+            lv_btnmatrix_clear_btn_ctrl(g_standardJumpButtonWidgets.inputAccountKeyboard, 11, LV_BTNMATRIX_CTRL_CHECKED);
+        }
+        lv_btnmatrix_set_btn_ctrl(g_standardJumpButtonWidgets.inputAccountKeyboard, 11, LV_BTNMATRIX_CTRL_DISABLED);
+    }
+}
 #endif
