+++ src/ui/gui_widgets/multi/web3/gui_connect_ada_widgets.c
@@ -1,4 +1,3 @@
-#ifndef BTC_ONLY
 #include "gui_connect_ada_widgets.h"
 #include "gui_connect_wallet_widgets.h"
 #include "gui.h"
@@ -130,5 +129,4 @@ static void OpenMoreHandler(lv_event_t *e)
                                     OpenTutorialHandler, &g_walletIndex);
     lv_obj_align(btn, LV_ALIGN_BOTTOM_MID, 0, -24);
 }
-#endif
